+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [next]
+## 4.7.4
 
 * fix: tap on `DatePicker` day does not skip a day ([#914](https://github.com/bdlukaa/fluent_ui/issues/914))
 * fix: ensure `PaneItemExpander`'s flyout is attached before using it ([#857](https://github.com/bdlukaa/fluent_ui/issues/857))